+++ routes/administration.js
@@ -3,9 +3,15 @@ const router   = express.Router()
 
 const pool = require('../config/pool')
 const { previewString, prettyDateTime } = require('../helpers/display')
+const { isAuth } = require('../helpers/auth')
 const { isAdmin } = require('../helpers/power')
 
-router.get('/', isAdmin, (req, res) => {
+router.get('/', isAuth, (req, res) => {
+    if(!isAdmin(req.user)) {
+        req.flash('error', `autorisations insuffisantes`)
+        return res.redirect('/actualites')
+    }
+
     let command = `SELECT COUNT(id) AS numPosts FROM posts;
     SELECT COUNT(id) AS numUsers FROM users WHERE power = ?;
     SELECT COUNT(id) AS numAdmins FROM users WHERE power > ?;
@@ -28,7 +34,12 @@ router.get('/', isAdmin, (req, res) => {
     })
 })
 
-router.get('/pending', isAdmin, (req, res) => {
+router.get('/pending', isAuth, (req, res) => {
+    if(!isAdmin(req.user)) {
+        req.flash('error', `autorisations insuffisantes`)
+        return res.redirect('/actualites')
+    }
+
     let command = `SELECT p.id, title, content, p.created_at, username
     FROM posts p JOIN users u ON p.user_id = u.id
     WHERE pending = true
@@ -52,7 +63,12 @@ router.get('/pending', isAdmin, (req, res) => {
     })
 })
 
-router.get('/pending/:id', isAdmin, (req, res) => {
+router.get('/pending/:id', isAuth, (req, res) => {
+    if(!isAdmin(req.user)) {
+        req.flash('error', `autorisations insuffisantes`)
+        return res.redirect('/actualites')
+    }
+
     let command = `SELECT p.id, title, content, p.created_at, p.user_id, username
     FROM posts p JOIN users u ON p.user_id = u.id WHERE p.id = ?;
     SELECT id, name FROM files WHERE post_id = ?;`
@@ -74,7 +90,12 @@ router.get('/pending/:id', isAdmin, (req, res) => {
     })
 })
 
-router.get('/pending/accepter/:id', isAdmin, (req, res) => {
+router.get('/pending/accepter/:id', isAuth, (req, res) => {
+    if(!isAdmin(req.user)) {
+        req.flash('error', `autorisations insuffisantes`)
+        return res.redirect('/actualites')
+    }
+
     let command = `UPDATE posts SET pending = false WHERE id = ?`
     let params = [req.params.id]
 
@@ -90,7 +111,12 @@ router.get('/pending/accepter/:id', isAdmin, (req, res) => {
     })
 })
 
-router.get('/pending/refuser/:id', isAdmin, (req, res) => {
+router.get('/pending/refuser/:id', isAuth, (req, res) => {
+    if(!isAdmin(req.user)) {
+        req.flash('error', `autorisations insuffisantes`)
+        return res.redirect('/actualites')
+    }
+
     let command = `DELETE FROM posts WHERE id = ?`
     let params = [req.params.id]
 
@@ -106,7 +132,12 @@ router.get('/pending/refuser/:id', isAdmin, (req, res) => {
     })
 })
 
-router.get('/utilisateurs', isAdmin, (req, res) => {
+router.get('/utilisateurs', isAuth, (req, res) => {
+    if(!isAdmin(req.user)) {
+        req.flash('error', `autorisations insuffisantes`)
+        return res.redirect('/actualites')
+    }
+
     let command = `SELECT id, username, power from users`
 
     pool.getConnection((error, connection) => {
@@ -122,7 +153,12 @@ router.get('/utilisateurs', isAdmin, (req, res) => {
     })
 })
 
-router.get('/infos', isAdmin, (req, res) => {
+router.get('/infos', isAuth, (req, res) => {
+    if(!isAdmin(req.user)) {
+        req.flash('error', `autorisations insuffisantes`)
+        return res.redirect('/actualites')
+    }
+    
     let command = `SELECT id, apropos, num_filieres, num_formateurs, num_stagiaires FROM infos;`
 
     pool.getConnection((error, connection) => {
